Further refactoring for the startup / shutdown overhaul.
